Fix up P_LookForPlayers().
It passes our test suite just as well as the the old function, but is much clearer and doesn't rely on direct access to the players vector.  I also think it covers more corner cases with gaps in playeringame, but I don't think we have a demo that covers that.

git-svn-id: http://odamex.net/svn/root/trunk@4120 b0bcef46-2027-0410-bd56-ab726e9a1d04